More reorganization. Add new column, "Is EMM". Try to add some issue tracking by auto creating issues in github when exceptions occur
