+++ Common/Globals.py
@@ -150,7 +150,7 @@ CSV_TAG_ATTR_NAME = {
     "Status": "Status",
     "Esper Version": "esper_client",
     "EEA Version": "eeaVersion",
-    # "EMM Device": "emm_device",
+    "Is EMM": "is_emm",
     "Template": "template_name",
     "Policy": "policy_name",
     "Mode": "Mode",
