+++ Utility/Logging/ApiToolLogging.py
@@ -5,12 +5,18 @@ import platform
 import sys
 import os
 import traceback
+
+from matplotlib.style import context
 import Common.Globals as Globals
 import Common.ApiTracker as ApiTracker
 import threading
 
 from datetime import datetime
 from traceback import print_exc, extract_tb, format_list
+
+from Utility.Logging.IssueTracker import IssueTracker
+
+from fuzzywuzzy import fuzz
 
 
 class ApiToolLog:
@@ -59,19 +65,25 @@ class ApiToolLog:
                 pass
 
     def LogError(self, e, exc_type=None, exc_value=None, exc_traceback=None):
-        if exc_type == None or exc_value == None or exc_traceback == None:
+        if exc_type is None or exc_value is None or exc_traceback is None:
             exc_type, exc_value, exc_traceback = sys.exc_info()
             exc_traceback = format_list(extract_tb(exc_traceback))
 
         self.limitLogFileSizes()
+        content = [
+            "\n%s\t: An Error has occured: %s\n" % (datetime.now(), e),
+            str(exc_type),
+            str(exc_value),
+        ]
+        for line in exc_traceback:
+            content.append(line)
 
         with open(self.logPath, "a") as myfile:
-            myfile.write("\n%s\t: An Error has occured: %s\n" % (datetime.now(), e))
-            myfile.write(str(exc_type))
-            myfile.write(str(exc_value))
-            for line in exc_traceback:
-                myfile.write(str(line))
+            for entry in content:
+                myfile.write(entry)
             myfile.write("\n")
+
+        self.postIssueToTrack(e, content)
 
     def LogPlace(self, str):
         with open(self.placePath, "a") as myfile:
@@ -82,23 +94,25 @@ class ApiToolLog:
             myfile.write(response)
 
     def excepthook(self, type, value, tb):
-        message = "\n%s\tUncaught exception:\n" % datetime.now()
-        print_exc()
         exc_type, exc_value, exc_traceback = sys.exc_info()
         exc_traceback = format_list(extract_tb(exc_traceback))
-        message += "".join(traceback.format_exception(type, value, tb))
-        message += "Exc Type: %s\n" % str(exc_type) if str(exc_type) else type
-        message += "Exc Value: %s\n" % str(exc_value) if str(exc_value) else value
-        message += (
+        content = [
+            "\n%s\tUncaught exception:\n" % datetime.now(),
+            "".join(traceback.format_exception(type, value, tb)),
+            "Exc Type: %s\n" % str(exc_type) if str(exc_type) else type,
+            "Exc Value: %s\n" % str(exc_value) if str(exc_value) else value,
             "Exc Traceback:\n%s\n" % str(exc_traceback) if str(exc_traceback) else tb
-        )
+        ]
+        print_exc()
         for line in exc_traceback:
-            message += str(line)
-        message += "\n"
-        print(message)
+            content.append(str(line))
         self.limitLogFileSizes()
         with open(self.logPath, "a") as myfile:
-            myfile.write(message)
+            for entry in content:
+                myfile.write(entry)
+            myfile.write("\n")
+
+        self.postIssueToTrack("%s %s" % (content[2], content[3]), content)
 
     def LogApiRequestOccurrence(self, src, api_func, writeToFile=False):
         if "main" in threading.current_thread().name.lower():
@@ -172,3 +186,34 @@ class ApiToolLog:
             finally:
                 if Globals.api_log_lock.locked():
                     Globals.api_log_lock.release()
+
+    def postIssueToTrack(self, excpt, content):
+        def getStrRatioSimilarity(s, t, usePartial=False):
+            if usePartial:
+                return fuzz.partial_ratio(s.lower(), t.lower())
+            return fuzz.ratio(s.lower(), t.lower())
+        tracker = IssueTracker()
+        title = None
+        if excpt is not None:
+            content.insert(0, str(excpt))
+        context.append("EAST Version:\t" % Globals.VERSION)
+        body = "\n".join(str(entry) for entry in content)
+
+        if isinstance(excpt, Exception):
+            title = repr(excpt)
+        elif excpt is not None:
+            title = str(excpt)
+
+        issues = tracker.listOpenIssues()
+        if issues:
+            match = False
+            for issue in issues:
+                ratio = getStrRatioSimilarity(issue["title"], title, True)
+                if ratio > 90:
+                    match = True
+                    tracker.postIssueComment(issue["number"], content)
+                    break
+            if not match:
+                tracker.createIssue(title, body)
+        else:
+            tracker.createIssue(title, body)
